CR-998 converted RetryableCloudDataStore to interface
