+++ src/main/java/uk/gov/ons/ctp/common/cloud/RetryableCloudDataStore.java
@@ -1,87 +1,65 @@
 package uk.gov.ons.ctp.common.cloud;
 
-import com.godaddy.logging.Logger;
-import com.godaddy.logging.LoggerFactory;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.retry.annotation.Backoff;
-import org.springframework.retry.annotation.Recover;
-import org.springframework.retry.annotation.Retryable;
-import org.springframework.stereotype.Service;
 import uk.gov.ons.ctp.common.error.CTPException;
-import uk.gov.ons.ctp.common.error.CTPException.Fault;
 
 /**
- * Decorator for <code>CloudDataStore</code>. It is responsible for handling exponential backoffs
- * when the datastore is becoming overloaded.
+ * Abstraction to a document data store in the cloud, with robust retry capability during store
+ * operation.
  */
-@Service
-public class RetryableCloudDataStore {
-  private static final Logger log = LoggerFactory.getLogger(RetryableCloudDataStore.class);
-
-  private CloudDataStore cloudDataStore;
-
-  @Autowired
-  public RetryableCloudDataStore(CloudDataStore cloudDataStore) {
-    this.cloudDataStore = cloudDataStore;
-    this.cloudDataStore.connect();
-  }
-
-  public void storeObject(
-      final String schema, final String key, final Object value, final String id)
-      throws CTPException {
-    try {
-      this.storeObject(schema, key, value);
-    } catch (DataStoreContentionException e) {
-      String identity = value.getClass().getName() + ": " + id;
-      log.error("Retries exhausted for storage of {}", identity);
-      throw new CTPException(Fault.SYSTEM_ERROR, e, "Retries exhausted for storage of " + identity);
-    }
-  }
-
-  @Retryable(
-      label = "storeObject",
-      include = DataStoreContentionException.class,
-      backoff =
-          @Backoff(
-              delayExpression = "#{${cloudStorage.backoffInitial}}",
-              multiplierExpression = "#{${cloudStorage.backoffMultiplier}}",
-              maxDelayExpression = "#{${cloudStorage.backoffMax}}"),
-      maxAttemptsExpression = "#{${cloudStorage.backoffMaxAttempts}}",
-      listeners = "cloudRetryListener")
-  public void storeObject(final String schema, final String key, final Object value)
-      throws CTPException, DataStoreContentionException {
-    cloudDataStore.storeObject(schema, key, value);
-  }
-
-  public <T> Optional<T> retrieveObject(Class<T> target, final String schema, final String key)
-      throws CTPException {
-    return cloudDataStore.retrieveObject(target, schema, key);
-  }
-
-  public <T> List<T> search(
-      Class<T> target, final String schema, String[] fieldPath, String searchValue)
-      throws CTPException {
-    return cloudDataStore.search(target, schema, fieldPath, searchValue);
-  }
-
-  public Set<String> getCollectionNames() {
-    return cloudDataStore.getCollectionNames();
-  }
+public interface RetryableCloudDataStore {
 
   /**
-   * When attempts to retry object storage have been exhausted this method is invoked and it can
-   * then throw the exception (triggering Rabbit retries). If this is not done then the message
-   * won't be eligible for another attempt or writing to the dead letter queue.
+   * Write object to cloud collection. If the collection already holds an object with the specified
+   * key then the contents of the value will be overwritten.
    *
-   * @param e is the final exception in the storeObject retries.
-   * @throws Exception the exception which caused the final attempt to fail.
+   * <p>The implementation will employ a retry strategy if contention errors are detected.
+   *
+   * @param schema the name of the collection that the object will be added to.
+   * @param key key for the object within the collection.
+   * @param value the object to be written to the collection.
+   * @param id a readable identifier for the object for error reporting. This may or may not be
+   *     different from the <code>key</code> parameter; for instance if key were a hash then a
+   *     different identifier may be better for the logged error/exception.
+   * @throws CTPException an error has occurred, that could not be rectified by the retry strategy.
    */
-  @Recover
-  public void doRecover(Exception e) throws Exception {
-    log.with(e.getMessage()).debug("Datastore recovery throwing exception");
-    throw e;
-  }
+  void storeObject(final String schema, final String key, final Object value, final String id)
+      throws CTPException;
+
+  /**
+   * Read an object.
+   *
+   * @param <T> The object type that results should be returned in.
+   * @param target the class of the object type that results should be returned in.
+   * @param schema the name of the collection which holds the object.
+   * @param key identifies the object within the collection.
+   * @return Optional containing the object if it was found.
+   */
+  <T> Optional<T> retrieveObject(Class<T> target, final String schema, final String key)
+      throws CTPException;
+
+  /**
+   * Runs an object search. This returns objects whose field is equal to the search value.
+   *
+   * @param <T> The object type that results should be returned in.
+   * @param target the class of the object type that results should be returned in.
+   * @param schema is the schema to search.
+   * @param fieldPathElements is an array of strings that describe the path to the search field. eg,
+   *     [ "case", "addresss", "postcode" ]
+   * @param searchValue is the value that the field must equal for it to be returned as a result.
+   * @return the list of results.
+   * @throws CTPException on error
+   */
+  <T> List<T> search(
+      Class<T> target, final String schema, String[] fieldPathElements, String searchValue)
+      throws CTPException;
+
+  /**
+   * Get the names of top level cloud collections.
+   *
+   * @return a Set with the names of the current collections.
+   */
+  Set<String> getCollectionNames();
 }
